Move entity classes inside entity package
